+++ packages/calcite-components/src/tests/commonTests/utils.ts
@@ -74,26 +74,3 @@ export function getTagOrHTMLWithBeforeContent(componentTestSetup: TagOrHTML | Ta
 export function propToAttr(name: string): string {
   return name.replace(/([a-z0-9])([A-Z])/g, "$1-$2").toLowerCase();
 }
-
-export function validateCaretIndex({
-  page,
-  componentTag,
-  nestedInputTypeSelector = "input",
-  position,
-}: {
-  page: E2EPage;
-  componentTag: string;
-  nestedInputTypeSelector?: "textarea" | "input";
-  position?: number;
-}): Promise<boolean> {
-  return page.evaluate(
-    (position, componentTag, nestedInputTypeSelector) => {
-      const element = document.querySelector(componentTag) as HTMLElement;
-      const el = element.shadowRoot.querySelector(nestedInputTypeSelector);
-      return el.selectionStart === (position !== undefined ? position : el.value.length);
-    },
-    position,
-    componentTag,
-    nestedInputTypeSelector,
-  );
-}
